+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [0.1.1] - 28-11-2024
 
 ### Added
+ 
+ - User CRUD endpoints and logic (create user, find, update user and update password)
+ - Tools module - ExcaptionApiAdvice handler
+ - org.passay:passay:1.6.6 and password strength check library
+
+## [0.1.1] - 28-11-2024
+
+### Added
 
  - Added API Data Transfer GHA tests